# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AyuCTMS",
    "description": "AyuCTMS \u2014 CRO / Sponsor Clinical Trial Management System Backend API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/portfolio/overview": {
      "get": {
        "tags": [
          "Portfolio / Derived Analytics"
        ],
        "summary": "Get Portfolio Overview",
        "description": "Computes all portfolio metrics dynamically from canonical database records.\nNo redundant counters are stored.",
        "operationId": "get_portfolio_overview_api_v1_portfolio_overview_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioOverviewResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/portfolio/alerts": {
      "get": {
        "tags": [
          "Portfolio / Derived Analytics"
        ],
        "summary": "Get Portfolio Alerts",
        "description": "Returns aggregated actionable operational alerts:\n- Approvals expiring within 30 days or already expired\n- Delayed milestones\n- Open SAEs requiring expedited review\n- Critical protocol deviations",
        "operationId": "get_portfolio_alerts_api_v1_portfolio_alerts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PortfolioAlertItem"
                  },
                  "type": "array",
                  "title": "Response Get Portfolio Alerts Api V1 Portfolio Alerts Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/portfolio/milestones/upcoming": {
      "get": {
        "tags": [
          "Portfolio / Derived Analytics"
        ],
        "summary": "Get Upcoming Milestones",
        "description": "Returns aggregated upcoming milestones across studies, ordered by planned date.",
        "operationId": "get_upcoming_milestones_api_v1_portfolio_milestones_upcoming_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UpcomingMilestoneItem"
                  },
                  "title": "Response Get Upcoming Milestones Api V1 Portfolio Milestones Upcoming Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolio/enrollment/trend": {
      "get": {
        "tags": [
          "Portfolio / Derived Analytics"
        ],
        "summary": "Get Enrollment Trend",
        "description": "Computes cumulative and discrete participant enrollment trend timeseries\nfrom canonical participant enrollment dates.",
        "operationId": "get_enrollment_trend_api_v1_portfolio_enrollment_trend_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "study_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Study Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnrollmentTrendPoint"
                  },
                  "title": "Response Get Enrollment Trend Api V1 Portfolio Enrollment Trend Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolio/studies/{study_id}/enrollment/by-site": {
      "get": {
        "tags": [
          "Portfolio / Derived Analytics"
        ],
        "summary": "Get Enrollment By Site",
        "description": "Computes recruitment progress, planned target vs actual enrolled count,\nbroken down per site for the specified study.",
        "operationId": "get_enrollment_by_site_api_v1_portfolio_studies__study_id__enrollment_by_site_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Study Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SiteEnrollmentItem"
                  },
                  "title": "Response Get Enrollment By Site Api V1 Portfolio Studies  Study Id  Enrollment By Site Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EnrollmentTrendPoint": {
        "properties": {
          "date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "enrolled_count": {
            "type": "integer",
            "title": "Enrolled Count"
          },
          "cumulative_count": {
            "type": "integer",
            "title": "Cumulative Count"
          }
        },
        "type": "object",
        "required": [
          "enrolled_count",
          "cumulative_count"
        ],
        "title": "EnrollmentTrendPoint"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PortfolioAlertItem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "study_id": {
            "type": "string",
            "title": "Study Id"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          }
        },
        "type": "object",
        "required": [
          "type",
          "severity",
          "title",
          "description",
          "study_id"
        ],
        "title": "PortfolioAlertItem"
      },
...
      "PortfolioOverviewResponse": {
        "properties": {
          "as_of_date": {
            "type": "string",
            "title": "As Of Date"
          },
          "studies": {
            "$ref": "#/components/schemas/PortfolioStudiesSummary"
          },
          "sites": {
            "$ref": "#/components/schemas/PortfolioSitesSummary"
          },
          "participants": {
            "$ref": "#/components/schemas/PortfolioParticipantsSummary"
          },
          "safety": {
            "$ref": "#/components/schemas/PortfolioSafetySummary"
          }
        },
        "type": "object",
        "required": [
          "as_of_date",
          "studies",
          "sites",
          "participants",
          "safety"
        ],
        "title": "PortfolioOverviewResponse"
      },
      "PortfolioParticipantsSummary": {
        "properties": {
          "total_screened": {
            "type": "integer",
            "title": "Total Screened"
          },
          "actual_enrolled": {
            "type": "integer",
            "title": "Actual Enrolled"
          },
          "active_in_treatment": {
            "type": "integer",
            "title": "Active In Treatment"
          },
          "screen_failures": {
            "type": "integer",
            "title": "Screen Failures"
          }
        },
        "type": "object",
        "required": [
          "total_screened",
          "actual_enrolled",
          "active_in_treatment",
          "screen_failures"
        ],
        "title": "PortfolioParticipantsSummary"
      },
      "PortfolioSafetySummary": {
        "properties": {
          "open_cases": {
            "type": "integer",
            "title": "Open Cases"
          },
          "open_serious_cases": {
            "type": "integer",
            "title": "Open Serious Cases"
          }
        },
        "type": "object",
        "required": [
          "open_cases",
          "open_serious_cases"
        ],
        "title": "PortfolioSafetySummary"
      },
      "PortfolioSitesSummary": {
        "properties": {
          "total_registered": {
            "type": "integer",
            "title": "Total Registered"
          },
          "currently_activated": {
            "type": "integer",
            "title": "Currently Activated"
          }
        },
        "type": "object",
        "required": [
          "total_registered",
          "currently_activated"
        ],
        "title": "PortfolioSitesSummary"
      },
      "PortfolioStudiesSummary": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "active": {
            "type": "integer",
            "title": "Active"
          },
          "planned": {
            "type": "integer",
            "title": "Planned"
          },
          "completed": {
            "type": "integer",
            "title": "Completed"
          },
          "delayed": {
            "type": "integer",
            "title": "Delayed"
          }
        },
        "type": "object",
        "required": [
          "total",
          "active",
          "planned",
          "completed",
          "delayed"
        ],
        "title": "PortfolioStudiesSummary"
      },
...
      "SiteEnrollmentItem": {
        "properties": {
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "site_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Code"
          },
          "site_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Name"
          },
          "recruitment_target": {
            "type": "integer",
            "title": "Recruitment Target"
          },
          "actual_enrolled": {
            "type": "integer",
            "title": "Actual Enrolled"
          },
          "recruitment_percentage": {
            "type": "number",
            "title": "Recruitment Percentage"
          },
          "activation_status": {
            "type": "string",
            "title": "Activation Status"
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "recruitment_target",
          "actual_enrolled",
          "recruitment_percentage",
          "activation_status"
        ],
        "title": "SiteEnrollmentItem"
      },
...
      "UpcomingMilestoneItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "study_id": {
            "type": "string",
            "title": "Study Id"
          },
          "study_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Study Code"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "planned_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned Date"
          },
          "actual_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Date"
          }
        },
        "type": "object",
        "required": [
          "id",
          "study_id",
          "title",
          "status"
        ],
        "title": "UpcomingMilestoneItem"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}